api/requirements.txt
requests
numpy==1.16.2
google-search-results-serpwow
setuptools
scikit-image
Flask
pymongo
gunicorn
tensorflow==1.14.0

